fix(retrieval): compute the listwise loss in fp32
The cross-encoder backbone emits bf16 logits, and the recipe divided them by the
temperature while still in bf16. At temperature 0.1 that magnifies them 10x onto a
coarse part of the bf16 grid -- at magnitude ~32 the representable spacing is 0.25 --
which distorts the softmax and every gradient flowing through it.

This was not theoretical. v60 failed to reproduce v50 on the ported code path: it
trained the same distance from the base model but in a different direction and scored
worse on BRIGHT and BCP. Every v60/v61 loss lands exactly on the bf16 grid (0.3086,
0.3926, 0.4883) where v50/v51's do not.

Up-casting before the division fixes it, and costs nothing at this tensor size. The
same change is applied to the validation path so train and val losses stay comparable.

+++ nemo_automodel/recipes/retrieval/train_cross_encoder.py
@@ -115,7 +115,13 @@ class TrainCrossEncoderRecipe(TrainBiEncoderRecipe):
             # temperature as the bi-encoder path does. Backbones that scale inside
             # forward (LlamaBidirectionalForSequenceClassification) apply model.temperature
             # there; the two compose, and both default to 1.0.
-            outputs.logits = outputs.logits.view(-1, self.train_n_passages) / self.temperature
+            #
+            # Up-cast to fp32 BEFORE the division. The backbone emits bf16 logits, and
+            # dividing by a small temperature magnifies them (10x at temperature=0.1),
+            # pushing them onto a coarse part of the bf16 grid -- at magnitude ~32 the
+            # spacing is 0.25, which distorts the softmax and every gradient through it.
+            # Staying in fp32 through cross_entropy costs nothing at this tensor size.
+            outputs.logits = outputs.logits.view(-1, self.train_n_passages).float() / self.temperature
             loss = F.cross_entropy(outputs.logits, labels)
 
             loss_buffer.append(loss.clone().detach())
@@ -149,7 +155,8 @@ class TrainCrossEncoderRecipe(TrainBiEncoderRecipe):
 
                     with autocast_ctx:
                         outputs = model(**batch, return_dict=True)
-                        logits = outputs.logits.view(-1, self.val_n_passages) / self.temperature
+                        # fp32 before the temperature division, as in the train step.
+                        logits = outputs.logits.view(-1, self.val_n_passages).float() / self.temperature
                         loss = F.cross_entropy(logits, labels)
 
                     loss_buffer.append(loss.clone().detach())
